Fix UtilityConnectionsService error

+++ src/Ymir.GeminiSync.Domain/SyncError.cs
@@ -4,7 +4,7 @@ public class SyncError
 {
     public int? PlaceNr { get; set; }
 
-    public int? AgreementId { get; set; }
+    public long? AgreementId { get; set; }
 
     public string Description { get; set; }
 }
